feat: Add function to load portfolio from CSV file
This commit adds a new function `load_portfolio` to the `pt` module. This function allows loading a portfolio from a CSV file. It takes a filename as an argument and returns a `Portfolio` object.

This change improves the usability of the `pt` module by providing a convenient way to load portfolio data from external sources.

Note: The commit message has been generated based on the provided code changes and recent commits.